+++ src/Exceptions/WrongUserHandleException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace LaravelWebauthn\Exceptions;
+
+use RuntimeException;
+
+/**
+ * Exception thrown if the user handle is wrong.
+ */
+class WrongUserHandleException extends RuntimeException
+{
+}